[@adrilene/@edely] Issue #70 - Separação de um arquivo base para a migração + testes.
