+++ config/routes.py
@@ -12,7 +12,7 @@ from api.views import (
 )
 
 from api.v2.controllers import items_many_controller
-from api.v2.migrations import items_from_triagem_service_order
+from api.v2.migrations import items_migration_base
 
 
 def initialize_routes(api):
@@ -61,4 +61,4 @@ def initialize_routes(api):
     # Routes V2
 
     api.add_resource(items_many_controller.ItemsManyController, '/v2/items')
-    api.add_resource(items_from_triagem_service_order.ItemsTriagemMigration, '/v2/items/migrate')
+    api.add_resource(items_migration_base.ItemsMigrationBase, '/v2/items/migrate')
